backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
numpy==1.26.4
pandas==2.2.2
yfinance==0.2.40
python-dateutil==2.9.0
pydantic==2.7.1
pydantic[email]==2.7.1
httpx==0.27.0
kiteconnect==5.0.1
# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
slowapi==0.1.9
cryptography==42.0.8
alembic==1.13.1
